start groups module

+++ controller/users.js
@@ -1,19 +1,39 @@
 /* 用户管理模块 */
 const md5=require('md5')
 const {secert}=require('../config')
-const {User,Role,Group,Menu,Permission,Element,Action,File,UserGroup,UserRole}=require('../db/mysql/models')
+const {Sequelize,User,Role,Menu,Permission,Element,Action,File,UserGroup,UserRole,Hospital,HospitalUser}=require('../db/mysql/models')
 const {encrypt,decrypt}=require('../utils/tokenVerify')
 //分页查询
 exports.find=async ctx=>{
+    let {Roles,Hospitals}=await userInfo(ctx)
+    let roles=Roles.map(itm=>{
+        return itm.roleEncode
+    })
+    let hospitals=Hospitals.map(itm=>{
+        return itm.id;
+    })
+    let include;
+    if(roles.includes('superadmin')){
+        include=[
+            {
+                model:Hospital,through:{attributes:[]}
+            }
+        ]
+    }else{
+        include=[
+            {
+                model:Hospital,through:{attributes:[]},where:{
+                    id:{[Sequelize.Op.in]:hospitals}
+                }
+            }
+        ]
+    }
     let {pageIndex=1,pageSize=10}=ctx.query
     pageIndex=Math.max(pageIndex,1)
     pageSize=Math.max(pageSize,10)
     let result=await User.findAndCountAll({
         attributes:{exclude:['password']},
-        include:[
-            {model:Role,through:{attributes:[]}},
-            {model:Group,through:{attributes:[]}}
-        ],
+        include,
         limit:pageSize,
         offset:pageIndex-1
     })
@@ -30,37 +50,36 @@ exports.createUser=async ctx=>{
         phone:{type:"string",format:/^0?(13|14|15|17|18)[0-9]{9}$/,required:false},
         password:{type:"string",min:6,required:false},
         status:{type:"boolean",required:false},
+        hospitalId:{type:"array",required:true,itemType:"int"}
     })
-    let {password}=ctx.request.body;
+    let {Roles}=await userInfo(ctx)
+    let roles=Roles.map(itm=>{
+        return itm.roleEncode
+    })
+    let {password,hospitalId=[]}=ctx.request.body;
     password=md5(`${secert}${password}`)
+    //创建的用户需要和机构关联起来 
+    //只有超级管理员和管理员有创建用户的权利，那么就针对这个关联
+    if(!roles.includes('superadmin') && hospitalId.length===0){
+        ctx.throw(422,'该权限下创建的用户关联的医院不能为空')
+        return
+    }
     let user=await User.create({...ctx.request.body,password,createdAt:new Date(),updatedAt:new Date()})
-    //创建的用户需要和角色关联起来 
-    //只有超级管理员和管理员有创建用户的权利，那么就针对这个关联角色
-    let {roles}=ctx.state.user;
-    if(roles.length===0) return;
-    if(roles.length===1 && roles[0].roleEncode==="superadmin"){
-        await UserRole.create({roleId:roles[0].id,userId:user.id,createdAt:new Date(),updatedAt:new Date()})
-    }else{
-        //有问题
-        for(let i=0;i<roles.length;i++){
-            await UserRole.create({roleId:roles[i].id,userId:user.id,createdAt:new Date(),updatedAt:new Date()})
-        }
+    for(let i=0;i<hospitalId.length;i++){
+        await HospitalUser.create({userId:user.id,hospitalId:hospitalId[i]})
     }
 
-    ctx.body={user}
+    ctx.body={message:"success",status:200}
 }
 //登录
 exports.login=async ctx=>{
     let {loginName,password}=ctx.request.body;
         let user=await User.findOne({
-            where:{loginName,password},
-            include:[
-                {model:Role,through:{attributes:[]},attributes:['roleName','roleEncode']}
-            ]
+            where:{loginName,password}
         })
         user=JSON.parse(JSON.stringify(user))
         if(!user) ctx.throw(401,'user not exist, login first')
-        let token=encrypt({loginName,id:user.id,roles:user.Roles},'1d')
+        let token=encrypt({loginName,id:user.id},'1d')
         ctx.body={token}
 }
 
@@ -71,8 +90,8 @@ exports.checkLoginStatus=async (ctx,next)=>{
         let token=ctx.header.authorization
         if(!token) ctx.throw(401,'鉴权失败，请求头中没有返回token值')
         token=token.replace('Bearer ','')
-        let {loginName,id,roles}=decrypt(token)
-        ctx.state.user={loginName,id,roles}
+        let {loginName,id}=decrypt(token)
+        ctx.state.user={loginName,id}
         await next()
     }else{
         await next()
@@ -81,19 +100,21 @@ exports.checkLoginStatus=async (ctx,next)=>{
 }
 
 //请求个人信息并返回当前用户配对的菜单项
-exports.getLoginUserInfo=async ctx=>{
+const userInfo=async ctx=>{
     let {id}=ctx.state.user;
     let user=await User.findByPk(id,{
         attributes:{exclude:['password']},
         include:[
-            {model:Role,through:{attributes:[]},include:[
-                {model:Permission,through:{attributes:[]}}
-            ]}
+            {model:Role,through:{attributes:[]}},
+            {model:Hospital,through:{attributes:[]}}
         ]
     })
     user=JSON.parse(JSON.stringify(user));
+    return user;
+}
+exports.getLoginUserInfo=async ctx=>{
+    let user=await userInfo(ctx)
     let roles=user.Roles;
-    delete user.Roles;
     let {menus=[],elements=[],files=[]}=await matchCurUserMenus(roles)
     ctx.body={user,menus,elements,files}
 }
@@ -101,12 +122,13 @@ exports.getLoginUserInfo=async ctx=>{
 //删改查findUserById,updateUserById,removeUserById
 exports.findUserById=async ctx=>{
     let {id}=ctx.params
-        ctx.body=(await User.findByPk(id,{
-            attributes:{exclude:['password']},
-            include:[
-                {model:Role,through:{attributes:[]}}
-            ]
-        }))
+    ctx.body=(await User.findByPk(id,{
+        attributes:{exclude:['password']},
+        include:[
+            {model:Role,through:{attributes:[]}},
+            {model:Hospital,through:{attributes:[]}}
+        ]
+    }))
 }
 exports.updateUserById=async ctx=>{
     let {id}=ctx.params
@@ -116,7 +138,8 @@ exports.updateUserById=async ctx=>{
             email:{type:"email",required:false},
             phone:{type:"string",format:/^0?(13|14|15|17|18)[0-9]{9}$/,required:false},
             password:{type:"string",min:6,required:false},
-            status:{type:"boolean",required:false}
+            status:{type:"boolean",required:false},
+            hospitalId:{type:"array",itemType:"int",required:true}
         })
         let {password}=ctx.request.body;
         password=md5(`${secert}${password}`)
@@ -134,36 +157,47 @@ exports.removeUserById=async ctx=>{
         ctx.throw(412,'用户拥有超级管理员角色，不能被删除')
         return 
     }
-    let [a,b,c]=await Promise.all([UserRole.destroy({where:{userId:id}}),UserGroup.destroy({where:{userId:id}}),User.destroy({where:{id}})])
-    ctx.body="succuss"
+    let [a,b,c,d]=await Promise.all([UserRole.destroy({where:{userId:id}}),UserGroup.destroy({where:{userId:id}}),User.destroy({where:{id}}),HospitalUser.destroy({where:{userId:id}})])
+    ctx.body={message:"success",status:200}
 }
-//根据用户的角色匹配用户的菜单
+/**
+ * 根据用户的角色匹配用户的菜单
+ */
 const matchCurUserMenus=async roles=>{
+    let rolesId=[];
+    for(let i=0;i<roles.length;i++){
+        rolesId.push(roles[i].id)
+    }
+    let permissions=await Permission.findAll({
+        include:[
+            {model:Role,through:{attributes:[]},where:{
+                id:{[Sequelize.Op.in]:rolesId}
+            }}
+        ]
+    })
+    permissions=JSON.parse(JSON.stringify(permissions))
     let userMenus=new Map(),
-            userElements=new Map(),
-            userFiles=new Map()
-        for(let i=0;i<roles.length;i++){
-            let permissions=roles[i].Permissions;
-            for(let i=0;i<permissions.length;i++){
-                if(permissions[i].typeName==='MENU'){
-                    let t_ms=userMenus.get(permissions[i].typeId)||[];
-                    userMenus.set(permissions[i].typeId,t_ms)
-                    if(!permissions[i].actionId) continue;
-                    userMenus.set(permissions[i].typeId,[...new Set([...t_ms,permissions[i].actionId])])
+        userElements=new Map(),
+        userFiles=new Map()
+        for(let i=0;i<permissions.length;i++){
+            if(permissions[i].typeName==='MENU'){
+                let t_ms=userMenus.get(permissions[i].typeId)||[];
+                userMenus.set(permissions[i].typeId,t_ms)
+                if(!permissions[i].actionId) continue;
+                userMenus.set(permissions[i].typeId,[...new Set([...t_ms,permissions[i].actionId])])
 
-                }
-                else if(permissions[i].typeName==='ELEMENT'){
-                    let t_ms=userElements.get(permissions[i].typeId)||[];
-                    userElements.set(permissions[i].typeId,t_ms)
-                    if(!permissions[i].actionId) continue;
-                    userElements.set(permissions[i].typeId,[...new Set([...t_ms,permissions[i].actionId])])
-                }
-                else if(permissions[i].typeName==="FILE"){
-                    let t_ms=userFiles.get(permissions[i].typeId)||[];
-                    userFiles.set(permissions[i].typeId,t_ms)
-                    if(!permissions[i].actionId) continue;
-                    userFiles.set(permissions[i].typeId,[...new Set([...t_ms,permissions[i].actionId])])
-                }
+            }
+            else if(permissions[i].typeName==='ELEMENT'){
+                let t_ms=userElements.get(permissions[i].typeId)||[];
+                userElements.set(permissions[i].typeId,t_ms)
+                if(!permissions[i].actionId) continue;
+                userElements.set(permissions[i].typeId,[...new Set([...t_ms,permissions[i].actionId])])
+            }
+            else if(permissions[i].typeName==="FILE"){
+                let t_ms=userFiles.get(permissions[i].typeId)||[];
+                userFiles.set(permissions[i].typeId,t_ms)
+                if(!permissions[i].actionId) continue;
+                userFiles.set(permissions[i].typeId,[...new Set([...t_ms,permissions[i].actionId])])
             }
         }
 
@@ -172,15 +206,12 @@ const matchCurUserMenus=async roles=>{
         
         return {menus,elements,files}
 }
-
+//根据类型查找对应范围内的数据值
 const findCurrentUserItems=async (type,aId)=>{
-    let types=await type.findAll()
-    let tmp=[]
-    types=JSON.parse(JSON.stringify(types))
-    for(let i=0;i<types.length;i++){
-        if(aId.includes(types[i].id)){
-            tmp.push(types[i]) 
+    let types=await type.findAll({
+        where:{
+            id:{[Sequelize.Op.in]:aId}
         }
-    }
-    return tmp;
+    })
+    return types;
 }
